Fix traces unmarshal error on fractional mem_delta (#21)
* Fix traces unmarshal error on fractional mem_delta

The Scout API reports a trace's mem_delta as a float in megabytes
(e.g. 0.0, 2.9296875), but TraceEntry and TraceDetail declared it as
int64, so `scout traces list` and `scout traces show` failed with:

  json: cannot unmarshal number 2.9296875 into Go struct field
  TraceEntry.traces.mem_delta of type int64

Decode mem_delta as float64 and format it as MB (matching the Scout UI)
via a new output.FormatMB helper, instead of treating it as a byte count.
Adds regression tests using live-shaped payloads (float, 0.0, null and
integer mem_delta values).

Closes #19

* Treat trace total_call_time as milliseconds

The API reports total_call_time in milliseconds (the Scout UI renders it
with an "ms" suffix and Request#total_call_time_in_seconds converts from
ms). `scout traces list` formatted it as seconds, so an 83-second request
displayed as "83483.9s". Use FormatMs for the Duration column and drop
the extra *1000 in the legacy-trace fallback of the span tree header.

+++ internal/api/types.go
@@ -86,7 +86,7 @@ type TraceEntry struct {
 	ID            int                    `json:"id"`
 	Time          string                 `json:"time"`
 	TotalCallTime float64                `json:"total_call_time"`
-	MemDelta      int64                  `json:"mem_delta"`
+	MemDelta      float64                `json:"mem_delta"` // memory increase in MB
 	MetricName    string                 `json:"metric_name"`
 	URI           string                 `json:"uri"`
 	Context       map[string]interface{} `json:"context"`
@@ -115,7 +115,7 @@ type TraceDetail struct {
 	ID                int         `json:"id"`
 	Time              string      `json:"time"`
 	TotalCallTime     float64     `json:"total_call_time"`
-	MemDelta          int64       `json:"mem_delta"`
+	MemDelta          float64     `json:"mem_delta"` // memory increase in MB
 	MetricName        string      `json:"metric_name"`
 	URI               string      `json:"uri"`
 	TransactionID     string      `json:"transaction_id"`
